# Fix compilation error when defining already inlined by compiler methods