Fixes https://github.com/vt-elixir/ja_serializer/issues/283
Related https://github.com/vt-elixir/ja_serializer/pull/245

Benchmark before the change:
```
Settings:
  duration:      1.0 s

## JaSerializer.PhoenixViewBench
[17:16:16] 1/6: attributes map big data
[17:16:21] 2/6: attributes map small data
[17:16:23] 3/6: render index 25 items map big data
[17:16:25] 4/6: render index 25 items small data
[17:16:29] 5/6: render index one item map big data
[17:16:32] 6/6: render index one item small data

Finished in 17.87 seconds

## JaSerializer.PhoenixViewBench
attributes map small data             10000000   0.12 µs/op
attributes map big data               10000000   0.46 µs/op
render index one item small data         20000   84.84 µs/op
render index one item map big data       20000   93.65 µs/op
render index 25 items small data         10000   292.75 µs/op
render index 25 items map big data        2000   820.56 µs/op
```

Benchmark after the change

```
Settings:
  duration:      1.0 s

## JaSerializer.PhoenixViewBench
[17:15:36] 1/6: attributes map big data
[17:15:42] 2/6: attributes map small data
[17:15:43] 3/6: render index 25 items map big data
[17:15:46] 4/6: render index 25 items small data
[17:15:48] 5/6: render index one item map big data
[17:15:50] 6/6: render index one item small data

Finished in 17.07 seconds

## JaSerializer.PhoenixViewBench
attributes map small data             10000000   0.13 µs/op
attributes map big data               10000000   0.53 µs/op
render index one item small data         20000   83.12 µs/op
render index one item map big data       20000   92.74 µs/op
render index 25 items small data          5000   288.62 µs/op
render index 25 items map big data        2000   785.51 µs/op
```

Best!