@beerlington 2019-01-07
@marpo60 I just tried running this on my production app and am getting errors related to overriding attributes in the view like this:

```
defmodule MyView do
  use JaSerializer.PhoenixView

  attributes([:my_attr])

  def my_attr(%{my_attr: my_attr}), do: my_attr
end
```

produces...
```
function MyView.my_attr/1 is undefined or private
```

I haven't had a chance to dig in too much, but maybe there isn't test coverage around this or my app is doing something we're not testing. I might need to revert this if I can't figure out what's going on.