@marpo60 2019-01-07
@beerlington I just try both cases on a local branch and they are working for me.
Im not able to reproduce the bug 😕

https://github.com/marpo60/ja_serializer/commit/2af4ee9a968a1e48ce4ba629d71e31ea1bb5318a