@beerlington 2019-01-08
@marpo60 I dug in a little more and think I created a test that reproduces it. I want to spend a little more time with it this evening and see if it's something we're doing in our production app or not.